discovery: relocate DiscoverySchemaChanged to errors.py
Move the exception next to the other BaseMngrError subclasses so the
canonical error registry stays centralized. Re-exported through
discovery_events so existing imports continue to work.
