+++ libs/mngr/imbue/mngr/api/discovery_events.py
@@ -29,6 +29,7 @@ from imbue.imbue_common.pure import pure
 from imbue.mngr.config.data_types import MngrConfig
 from imbue.mngr.config.data_types import MngrContext
 from imbue.mngr.errors import BaseMngrError
+from imbue.mngr.errors import DiscoverySchemaChanged
 from imbue.mngr.interfaces.data_types import AgentDetails
 from imbue.mngr.interfaces.host import OnlineHostInterface
 from imbue.mngr.primitives import AgentId
@@ -42,22 +43,6 @@ from imbue.mngr.primitives import ProviderInstanceName
 from imbue.mngr.primitives import SSHInfo
 
 DISCOVERY_EVENT_SOURCE: Final[EventSource] = EventSource("mngr/discovery")
-
-
-class DiscoverySchemaChanged(BaseMngrError, ValueError):
-    """Raised when a discovery event line cannot be validated against the current schema.
-
-    This typically means a field was added, removed, or renamed in a discovery event
-    model since the line was written. Callers should treat the on-disk events as stale,
-    regenerate via a full discovery (which appends new events in the current schema),
-    and retry. If validation fails again after regeneration, the error is real and
-    should be surfaced rather than silently dropped.
-    """
-
-    def __init__(self, event_type: str, validation_error: str) -> None:
-        self.event_type = event_type
-        self.validation_error = validation_error
-        super().__init__(f"Discovery event of type {event_type!r} does not match current schema: {validation_error}")
 
 
 class DiscoveryEventType(UpperCaseStrEnum):
